Refactor applicants tab and restructure folders
